Using OpenApi Generator to create the ProfileApi and the associated models
We want to avoid the manual creation of the APIs and the model classes. The best way to do this is to define the public API following the OpenApi format in a .yaml file.

At the moment, Gravatar doesn't have those definitions, but we want to be ready, so we are adding the ProfileApi definition that we've created. (Obviously, it's the first version so it can contain errors or lack information).

We are using OpenApi Generator - https://openapi-generator.tech

As we want to have the APIs and the models within the project, we've modified the Gradle task in order to generate all the files in "build" folder and move only the desired files to the project directories.

The generated files didn't look exactly as we wanted, so we'll modify some of the OpenApi generator templates to match what we want.

The last step in the generation process consist on formatting the code using Ktlint.