Implement GravatarTheme
It would be nice to have our own GravatarTheme so that we can keep our Gravatar style in all UI components, at least for the default state.

Developers using the SDK could override our Theme if they don't want to use it. Using CompositionLocal, they'll be able to provide another implementation of GravatarTheme using whatever they want.

Right now, we are using the Material 3 base styles, but we plan to update those with our colors, typographies, and shapes.