targetSdk=34 in `gravatar-ui`
The user of targetSdk in a library module is deprecated. However, using minSdk makes the composable preview launch fail with the following error:

"The application could not be installed: INSTALL_FAILED_DEPRECATED_SDK_VERSION"

I haven't found a suitable solution for the issue, but Google is using that approach in its Accompanist repo.

https://github.com/google/accompanist/blob/879ec1a2c937871ab005c7ae143c135bf29f51f1/insets-ui/build.gradle.kts#L37C9-L38C23

We probably need to review this in the future, but it looks like a valid workaround at the moment.