Add Gravatar SDK initialisation method
Clients should use the API key generated in the Gravatar Developer portal to get complete information from the Gravatar SDK.

When you don't use the API Key you'll get the basic information. That info is the minimum needed to show the Profile components in the UI module.

We are providing a method to initialise the SDK with the API Key. After that, all the requests will use that key.