Map DateTime format strings to `java.time.Instant
Instant class is a better representation of the DateTime format in our SDK.

We also need to add an adapter to deserialize and string with the proper format to an Instant instance.

Some `java.time` classes are not available by default on API < 26. For that reason, we need to add desugaring on the core module.