Non-catching methods throw exceptions for all kind of errors
For handling errors, we were returning a Result<Profile, ErrorType> and throwing exceptions at the same time. That will force developers to handle errors in two different ways, looking at the ErrorType and catching possible exceptions.

With these changes, we throw an exception for all errors so developers will receive a Profile or an exception.