Update MediaPickerPopop logic to be globally positioned
The positioning of the poped proved to be tricky with the previous logic as the popup would be placed only in the Parent composable. This is not enough for the vertical scroll where the upload button can be barely visible.
This implemetation draws a full screen transparent dialog and the popup on top of it so we can position the popup on the whole screen.