Avatar composable with forceRefresh option
When the avatar is updated, we need to refresh the current image. For that purpose, we need the cache buster in the request.

On the composable side, as the state does not change (because the avatar URL is the same), we need to force the recomposition. In this case, we use the "key" with the selected image ID.