Refreshing all info when no avatar selected and a new one is uploaded
When a new avatar is uploaded into a Gravatar account with no avatar selected, the backend automatically uses that image as the default avatar.

We need to reflect that in the QE UI. We can hardcode the changes in the code, modifying the uiState as required. However, if the backend behavior changes at some point, all previous versions of the QE will show the wrong status of the Gravatar account.

To mitigate that situation, if an image is uploaded and no previous avatar is selected, we will refresh all data from the backend.