Added a new ErrorType to handle possible future errors
That will force any developer to add that consideration in their current exhaustive when clauses.

During the `2.x.x` every new ErrorType that we want to add should inherit from `FutureErrors`. Any exhaustive code will still compile as they are already considering the parent class. If someone is interested in a new concrete error, they can add that branch to the when without problems while keeping the FutureErrors branch.

When we release new major versions, we need to change the parent class for the new errors:

The exact process will be followed during the life of a major version.